+++ azext_iot/_help.py
@@ -1042,6 +1042,18 @@ helps[
 """
 
 helps[
+    "iot edge export-modules"
+] = """
+    type: command
+    short-summary: Export the edge modules' configuration on a single edge device.
+    long-summary: The module twin configuration output can be directly used as the --content of "az iot edge set-modules".
+    examples:
+    - name: Export module twin configuration on a target device.
+      text: >
+        az iot edge export-modules --hub-name {iothub_name} --device-id {device_id}
+"""
+
+helps[
     "iot edge deployment"
 ] = """
     type: group