+++ azext_iot/tests/utility/test_iot_utility_unit.py
@@ -477,3 +477,36 @@ class TestHandleServiceException(object):
 
         with pytest.raises(expected_error):
             handle_service_exception(error)
+
+
+class TestFileHeaders(object):
+    def test_file_headers(self):
+        from azext_iot.constants import EXTENSION_ROOT
+
+        header = """# coding=utf-8
+# --------------------------------------------------------------------------------------------
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License. See License.txt in the project root for license information.
+# --------------------------------------------------------------------------------------------"""
+
+        files_missing_header = []
+        sdk_root = "sdk"
+
+        def _validate_directory(path):
+            for entry in os.scandir(path):
+                if entry.is_dir(follow_symlinks=False) and sdk_root not in entry.path:
+                    _validate_directory(entry.path)
+                else:
+                    if entry.is_file() and entry.path.endswith(".py"):
+                        contents = None
+                        with open(entry.path, "rt", encoding="utf-8") as f:
+                            contents = f.read()
+                        if contents and not contents.startswith(header):
+                            files_missing_header.append(entry.path)
+
+        _validate_directory(EXTENSION_ROOT)
+        if files_missing_header:
+            pytest.fail(
+                "The following files are missing an encoding and license header, or it is improperly formatted:\n"
+                "{}".format("\n".join(files_missing_header))
+            )